Somewhat fix remaining tests
